requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
Django==1.9
djangorestframework==3.3.2

pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
six==1.4.1
xattr==0.6.4